Simplify update checking
Previously update checking was done via proxy with the react useEffect
hook. This was causing problems when the usePrevious hook was called but
its own props had not been changed; it would only supply the previous
value on the first call, after that it would supply the current value.

Now there is a conditional block that only runs if the target value has
changed - that updates the ref to use the previous value and remembers
the current value to check for future updates.